lesson_19/requirements.txt
allure-pytest==2.10.0
allure-python-commons==2.10.0
astunparse==1.6.3
async-generator==1.10
attrs==23.1.0
backports.functools-lru-cache==1.6.6
beautifulsoup4==4.12.2
blinker==1.6.2
Brotli==1.0.9
bs4==0.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
ci-info==0.3.0
click==8.1.3
configobj==5.0.8
configparser==5.3.0
cryptography==41.0.1
cvxpy==1.3.2
daytime==0.4
docutils==0.20.1
easyparse==1.0.6
ecos==2.0.12
enum34==1.1.10
etelemetry==0.3.0
exceptiongroup==1.1.2
filelock==3.12.2
fitz==0.0.1.dev2
Flask==2.3.2
future==0.18.3
gevent==23.7.0
googleapis-common-protos==1.59.0
greenlet==2.0.2
grpcio==1.54.0
h11==0.14.0
homepage==0.5.7
httplib2==0.22.0
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
looseversion==1.2.0
lxml==4.9.2
magiclog==1.0.2
mapping==0.1.6
MarkupSafe==2.1.3
mutagen==1.46.0
namedlist==1.8
networkx==3.1
nibabel==5.1.0
nipype==1.8.6
numpy==1.25.0
osqp==0.6.3
outcome==1.2.0
packaging==23.1
pages==0.3
pandas==2.0.2
pathlib==1.0.1
pdfminer.six==20221105
piglet-templates==1.3.0
playwright==1.36.0
pluggy==1.0.0
protobuf==4.22.3
prov==2.0.0
pycparser==2.21
pycryptodomex==3.18.0
pydot==1.4.2
pyee==9.0.4
PyMuPDF==1.22.5
pyparsing==3.1.0
PyPDF2==3.0.1
PyPDF4==1.27.0
PySocks==1.7.1
pytest==7.4.0
pytest-allure-adaptor==1.7.10
pytest-base-url==2.0.0
pytest-playwright==0.3.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2023.3
pyxnat==1.5
qdldl==0.1.7.post0
rdflib==6.3.2
requests==2.31.0
scipy==1.10.1
scs==3.2.3
selene==1.0.2
selenium==4.10.0
sequence==0.3.4
simplejson==3.19.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
stackclimber==0.99
text-unidecode==1.3
tqdm==4.65.0
traits==6.3.2
trio==0.22.1
trio-websocket==0.10.3
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.2
utils==1.0.1
v1==0.12.3
v2==1.0.7
webdriver-manager==3.8.6
websockets==11.0.3
Werkzeug==2.3.6
wsproto==1.2.0
yt-dlp==2023.7.6
zipp==3.16.2
zope.event==5.0
zope.interface==6.0
